Reduce internal states of Block
