+++ x-pack/plugin/esql/compute/src/main/generated-src/org/elasticsearch/compute/data/BooleanArrayBlock.java
@@ -35,20 +35,18 @@ final class BooleanArrayBlock extends AbstractArrayBlock implements BooleanBlock
             positionCount,
             firstValueIndexes,
             nulls,
-            mvOrdering,
-            blockFactory
+            mvOrdering
         );
     }
 
     private BooleanArrayBlock(
-        BooleanArrayVector vector,
+        BooleanArrayVector vector, // stylecheck
         int positionCount,
         int[] firstValueIndexes,
         BitSet nulls,
-        MvOrdering mvOrdering,
-        BlockFactory blockFactory
+        MvOrdering mvOrdering
     ) {
-        super(positionCount, firstValueIndexes, nulls, mvOrdering, blockFactory);
+        super(positionCount, firstValueIndexes, nulls, mvOrdering);
         this.vector = vector;
         assert firstValueIndexes == null
             ? vector.getPositionCount() == getPositionCount()
@@ -115,8 +113,7 @@ final class BooleanArrayBlock extends AbstractArrayBlock implements BooleanBlock
             expandedPositionCount,
             null,
             shiftNullsToExpandedPositions(),
-            MvOrdering.DEDUPLICATED_AND_SORTED_ASCENDING,
-            blockFactory()
+            MvOrdering.DEDUPLICATED_AND_SORTED_ASCENDING
         );
         blockFactory().adjustBreaker(expanded.ramBytesUsedOnlyBlock() - bitSetRamUsedEstimate);
         // We need to incRef after adjusting any breakers, otherwise we might leak the vector if the breaker trips.
@@ -160,8 +157,12 @@ final class BooleanArrayBlock extends AbstractArrayBlock implements BooleanBlock
 
     @Override
     public void allowPassingToDifferentDriver() {
-        super.allowPassingToDifferentDriver();
         vector.allowPassingToDifferentDriver();
+    }
+
+    @Override
+    public BlockFactory blockFactory() {
+        return vector.blockFactory();
     }
 
     @Override
