+++ x-pack/plugin/esql/compute/src/main/generated-src/org/elasticsearch/compute/data/DoubleBigArrayBlock.java
@@ -36,20 +36,18 @@ public final class DoubleBigArrayBlock extends AbstractArrayBlock implements Dou
             positionCount,
             firstValueIndexes,
             nulls,
-            mvOrdering,
-            blockFactory
+            mvOrdering
         );
     }
 
     private DoubleBigArrayBlock(
-        DoubleBigArrayVector vector,
+        DoubleBigArrayVector vector, // stylecheck
         int positionCount,
         int[] firstValueIndexes,
         BitSet nulls,
-        MvOrdering mvOrdering,
-        BlockFactory blockFactory
+        MvOrdering mvOrdering
     ) {
-        super(positionCount, firstValueIndexes, nulls, mvOrdering, blockFactory);
+        super(positionCount, firstValueIndexes, nulls, mvOrdering);
         this.vector = vector;
         assert firstValueIndexes == null
             ? vector.getPositionCount() == getPositionCount()
@@ -116,8 +114,7 @@ public final class DoubleBigArrayBlock extends AbstractArrayBlock implements Dou
             expandedPositionCount,
             null,
             shiftNullsToExpandedPositions(),
-            MvOrdering.DEDUPLICATED_AND_SORTED_ASCENDING,
-            blockFactory()
+            MvOrdering.DEDUPLICATED_AND_SORTED_ASCENDING
         );
         blockFactory().adjustBreaker(expanded.ramBytesUsedOnlyBlock() - bitSetRamUsedEstimate);
         // We need to incRef after adjusting any breakers, otherwise we might leak the vector if the breaker trips.
@@ -161,8 +158,12 @@ public final class DoubleBigArrayBlock extends AbstractArrayBlock implements Dou
 
     @Override
     public void allowPassingToDifferentDriver() {
-        super.allowPassingToDifferentDriver();
         vector.allowPassingToDifferentDriver();
+    }
+
+    @Override
+    public BlockFactory blockFactory() {
+        return vector.blockFactory();
     }
 
     @Override
