+++ x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/data/AbstractVector.java
@@ -43,9 +43,4 @@ abstract class AbstractVector extends AbstractNonThreadSafeRefCounted implements
     protected void closeInternal() {
         blockFactory.adjustBreaker(-ramBytesUsed());
     }
-
-    @Override
-    public final boolean isReleased() {
-        return hasReferences() == false;
-    }
 }
