+++ x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/data/ConstantNullBlock.java
@@ -18,12 +18,21 @@ import java.util.Objects;
 /**
  * Block implementation representing a constant null value.
  */
-final class ConstantNullBlock extends AbstractBlock implements BooleanBlock, IntBlock, LongBlock, DoubleBlock, BytesRefBlock {
+final class ConstantNullBlock extends AbstractNonThreadSafeRefCounted
+    implements
+        BooleanBlock,
+        IntBlock,
+        LongBlock,
+        DoubleBlock,
+        BytesRefBlock {
 
     private static final long BASE_RAM_BYTES_USED = RamUsageEstimator.shallowSizeOfInstance(ConstantNullBlock.class);
+    private final int positionCount;
+    private BlockFactory blockFactory;
 
     ConstantNullBlock(int positionCount, BlockFactory blockFactory) {
-        super(positionCount, blockFactory);
+        this.positionCount = positionCount;
+        this.blockFactory = blockFactory;
     }
 
     @Override
@@ -221,5 +230,35 @@ final class ConstantNullBlock extends AbstractBlock implements BooleanBlock, Int
     public long getLong(int valueIndex) {
         assert false : "null block";
         throw new UnsupportedOperationException("null block");
+    }
+
+    @Override
+    public int getTotalValueCount() {
+        return 0;
+    }
+
+    @Override
+    public int getPositionCount() {
+        return positionCount;
+    }
+
+    @Override
+    public int getFirstValueIndex(int position) {
+        return 0;
+    }
+
+    @Override
+    public int getValueCount(int position) {
+        return 0;
+    }
+
+    @Override
+    public BlockFactory blockFactory() {
+        return blockFactory;
+    }
+
+    @Override
+    public void allowPassingToDifferentDriver() {
+        blockFactory = blockFactory.parent();
     }
 }
