+++ x-pack/plugin/esql/compute/src/main/java/org/elasticsearch/compute/data/DocBlock.java
@@ -20,7 +20,6 @@ public class DocBlock extends AbstractVectorBlock implements Block {
     private final DocVector vector;
 
     DocBlock(DocVector vector) {
-        super(vector.getPositionCount(), vector.blockFactory());
         this.vector = vector;
     }
 
@@ -189,5 +188,15 @@ public class DocBlock extends AbstractVectorBlock implements Block {
     @Override
     public void allowPassingToDifferentDriver() {
         vector.allowPassingToDifferentDriver();
+    }
+
+    @Override
+    public int getPositionCount() {
+        return vector.getPositionCount();
+    }
+
+    @Override
+    public BlockFactory blockFactory() {
+        return vector.blockFactory();
     }
 }
